controller/requirements.txt
PySide==1.2.2
msgpack-python==0.4.6
docopt==0.6.2
requests==2.7.0
requests-toolbelt==0.4.0
matplotlib==1.4.3
numpy==1.9.2
Flask==0.10.1
Flask-HTTPAuth==2.5.0
psycopg2==2.6
